requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aioredis==1.3.1
aiosignal==1.3.2
alpha_vantage==3.0.0
alpaca-py==0.10.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
backtrader==1.9.78.123
beautifulsoup4==4.13.4
black==25.1.0
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
channels==3.0.5
channels-redis==3.4.1
charset-normalizer==3.4.1
click==8.1.8
constantly==23.10.4
contourpy==1.3.2
coverage==7.8.0
cryptography==44.0.2
cycler==0.12.1
daphne==3.0.2
distlib==0.3.9
Django==5.2
django-filter==21.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dj-database-url>=2.0
filelock==3.18.0
flake8==7.2.0
fonttools==4.57.0
frozendict==2.4.6
frozenlist==1.6.0
google-generativeai==0.3.1
hiredis==3.1.0
hyperlink==21.0.0
identify==2.6.10
idna==3.10
incremental==24.7.2
iniconfig==2.1.0
isort==6.0.1
kiwisolver==1.4.8
matplotlib==3.10.1
mccabe==0.7.0
msgpack==1.1.0
multidict==6.4.3
multitasking==0.0.11
mypy==1.15.0
mypy-extensions==1.0.0
nodeenv==1.9.1
numpy==1.26.4
packaging==25.0
pandas<3.0.0
pandas-ta==0.3.14b0
pathspec==0.12.1
peewee==3.17.9
pillow==11.2.1
platformdirs==4.3.7
plotly==5.24.1
pluggy==1.5.0
pre_commit==4.2.0
propcache==0.3.1
protobuf==4.24.4
psycopg2-binary==2.9.10
py==1.11.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.13.0
pycparser==2.22
pyflakes==3.3.2
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyparsing==3.2.3
pytest==6.2.5
pytest-asyncio==0.20.3
pytest-cov==2.12.1
pytest-django==4.5.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
redis==4.6.0
requests==2.32.3
service-identity==24.2.0
setuptools==75.8.0
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
toml==0.10.2
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
urllib3==2.4.0
virtualenv==20.30.0
wheel==0.45.1
yarl==1.20.0
yfinance==0.2.56
zope.interface==7.2
